rearrange and combine files for ProgressStages
